+++ src/main/java/com/youzan/nsq/client/core/ConnectionManager.java
@@ -249,7 +249,7 @@ public class ConnectionManager {
                 try {
                     latch.await(RDY_TIMEOUT, TimeUnit.MILLISECONDS);
                 } catch (InterruptedException e) {
-                    logger.error("Interrupted waiting for rdy update for subscribe, topic {}, Connection {}", topic, subscriber);
+                    logger.warn("Interrupted waiting for rdy update for subscribe, topic {}, Connection {}", topic, subscriber);
                     Thread.currentThread().interrupt();
                 }
             } else {
@@ -314,16 +314,16 @@ public class ConnectionManager {
                 });
                 try {
                     if (!backoffLatch.await(RDY_TIMEOUT, TimeUnit.MILLISECONDS)) {
-                        logger.error("Timeout backoff topic connection {}", conn);
+                        logger.warn("Timeout backoff topic connection {}", conn);
                     } else if (null != latch) {
                         latch.countDown();
                     }
                 } catch (InterruptedException e) {
-                    logger.error("Interrupted waiting for rdy update for backoff, connection {}", conn);
+                    logger.warn("Interrupted waiting for rdy update for backoff, connection {}", conn);
                     Thread.currentThread().interrupt();
                 }
             } else {
-                logger.error("Connection {} does not belong to current consumer.", conn);
+                logger.warn("Connection {} does not belong to current consumer.", conn);
             }
         }
     }
@@ -370,13 +370,13 @@ public class ConnectionManager {
 
                 try{
                     if (!backoffLatch.await(latchCount * RDY_TIMEOUT, TimeUnit.MILLISECONDS)) {
-                        logger.error("Timeout backoff topic connections {}", topic);
+                        logger.warn("Timeout backoff topic connections {}", topic);
                     } else if (null != latch) {
                         latch.countDown();
                         logger.info("Backoff connections for topic {}", topic);
                     }
                 } catch (InterruptedException e) {
-                    logger.error("Interrupted while waiting for back off on all connections for {}", topic);
+                    logger.warn("Interrupted while waiting for back off on all connections for {}", topic);
                     Thread.currentThread().interrupt();
                 }
             } finally {
@@ -421,13 +421,13 @@ public class ConnectionManager {
 
                 try {
                     if (!resumeLatch.await(latchCount * RDY_TIMEOUT, TimeUnit.MILLISECONDS)) {
-                        logger.error("Timeout for resume topic connections {}", topic);
+                        logger.warn("Timeout for resume topic connections {}", topic);
                     } else if (null != latch) {
                         latch.countDown();
                         logger.info("Resume connections for topic {}", topic);
                     }
                 } catch (InterruptedException e) {
-                    logger.error("Interrupted while waiting for resume on all connections for {}", topic);
+                    logger.warn("Interrupted while waiting for resume on all connections for {}", topic);
                     Thread.currentThread().interrupt();
                 }
             } finally {
@@ -496,10 +496,10 @@ public class ConnectionManager {
                    //await for may rdy updates
                    try {
                        if (!latch.await(conCount * RDY_TIMEOUT, TimeUnit.MILLISECONDS)) {
-                           logger.error("Timeout for redistribute connections rdy {}", topic);
+                           logger.warn("Timeout for redistribute connections rdy {}", topic);
                        }
                    } catch (InterruptedException e) {
-                       logger.error("Interrupted while waiting for resume on all connections for {}", topic);
+                       logger.warn("Interrupted while waiting for resume on all connections for {}", topic);
                        Thread.currentThread().interrupt();
                    }
                } finally {