+++ src/test/java/it/youzan/nsq/client/ITProducer.java
@@ -118,6 +118,7 @@ public class ITProducer {
         }finally {
             stop.set(true);
             TopicUtil.deleteTopic(dailyLookupAdmin, topic);
+            producer.close();
             logger.info("[testLookupAddressUpdate] ends.");
         }
     }